+++ src/main/resources/templates/settings/entries/add.html
