+++ src/main/resources/templates/settings/entries/edit.html
