+++ src/main/resources/templates/settings/entries/index.html
